requirements.txt
python-telegram-bot==22.0
sqlite3